backend/requirements.txt
# Core Web Framework
fastapi
uvicorn[standard]
pydantic>=2.0 # Pydantic v2 is often used with FastAPI
python-dotenv
python-multipart # For FastAPI file uploads
protobuf>=5.29.0,<6.0.0 # Pin compatible Protobuf version

# LangChain Core & Community
langchain
langchain_core
langchain_community

# LLM Providers
langchain_groq # For Groq LLM
langchain-ollama # Preferred for Ollama Embeddings

# Vector Store - Specific versions to avoid circular imports
weaviate-client
langchain-weaviate

# Loaders
unstructured[pdf,docx,md,eml,html,xml] # For document parsing
python-magic-bin # Required by unstructured for file type detection on Windows
pypdf2
# Embeddings (Ollama is now primary, Sentence Transformers might be needed for future rerankers)
# sentence-transformers # Keep if planning to use certain rerankers, otherwise removable
torch # Often a dependency for sentence-transformers or faiss

# Image Handling (Dependency for Unstructured or potential future use)
Pillow

# Reranker (Optional, currently disabled and potentially problematic)
# flashrank

# OCR (Optional, only if processing scanned images via Tesseract)
# pytesseract

# Database
pymongo
motor # Async MongoDB driver
beanie # ODM for MongoDB with FastAPI integration